+++ packages/terra-draw/src/terra-draw.extensions.spec.ts
@@ -132,7 +132,6 @@ export class TerraDrawTestAdapter extends TerraDrawBaseAdapter {
 		_:
 			| "move"
 			| "unset"
-			| "default"
 			| "grab"
 			| "grabbing"
 			| "crosshair"
@@ -170,9 +169,8 @@ export class TerraDrawTestAdapter extends TerraDrawBaseAdapter {
  * custom draw modes for Terra Draw exclusively relying on the public API of the library.
  */
 
-interface TerraDrawTestModeOptions<
-	T extends CustomStyling,
-> extends TerraDrawExtend.BaseModeOptions<T> {
+interface TerraDrawTestModeOptions<T extends CustomStyling>
+	extends TerraDrawExtend.BaseModeOptions<T> {
 	customProperty: string;
 }