+++ packages/terra-draw-maplibre-gl-adapter/CHANGELOG.md
@@ -9,6 +9,12 @@ All notable changes to this project will be documented in this file. See [commit
 
 * **terra-draw-maplibre-gl-adapter:** add onCursorChange callback to delegate cursor control ([d6b32f7](https://github.com/blocktype-co-uk/terra-draw/commit/d6b32f7e24cdae3192c64258e9afb31a15e77370))
 
+
+### refactor
+
+* **terra-draw-maplibre-gl-adapter:** remove initial cursor preservation on unset (superseded by onCursorChange)
+* **terra-draw:** remove `"default"` from `SetCursor` union (superseded by onCursorChange)
+
 ## [1.4.1](https://github.com/JamesLMilner/terra-draw/compare/terra-draw-maplibre-gl-adapter@1.4.0...terra-draw-maplibre-gl-adapter@1.4.1) (2026-05-17)